requirements.txt
langchain==0.1.0
langgraph==0.0.20
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
psycopg2-binary==2.9.7
pgvector==0.1.8
SQLAlchemy==2.0.29
openai==1.3.0
sentence-transformers==2.2.2
reportlab==4.0.4
streamlit==1.28.0
fastapi==0.110.0
uvicorn[standard]==0.27.0
pytest==7.4.0
jupyter==1.0.0
